move test cases to components module
