+++ python_modules/libraries/dagster-dg/tox.ini
@@ -18,6 +18,7 @@ deps =
 allowlist_externals =
   /bin/bash
   uv
+  jsonschema
 commands =
   !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster'
   pytest ./dagster_dg_tests -vv {posargs}
